Bug Fix: Episode count is 0 fix
When the episode count is unknown by AniList, it is set to 0, this
causes a bug in the AnimePanel where the maximum and minimum value of
the spinner is set to 0 and you can't update your progress of the show.
This changes the 0 maximum to 255 so that proper progress may be made.

To do: make sure the new maximum value is set to something more
appropriate if available.