Temporary fix for crash
Context: All anime data is deleted when the user is deleted but the
anime data loads on a different interval than the user data so it causes
crashes between the two. I'll update it to reload the user without
deleting it eventually, for now it just reloads both on the same
interval.